+++ lib/resque/plugins/aps/application.rb
@@ -177,6 +177,63 @@ module Resque
           logger.error("ResqueAps[read_failed]: Bad data on the socket (#{name})") if logger
         end
 
+
+
+
+
+        def queued_count
+          (redis.get(Resque.aps_application_queued_key(name)) || 0).to_i
+        end
+
+        def lock_key
+          @lock_key ||= "#{Resque.aps_application_key(name)}:lock"
+        end
+        
+        def redis
+          Resque.redis
+        end
+        
+        def acquire_lock
+          if redis.setnx(lock_key, Time.now.utc.to_i + 1)
+            true
+          elsif Time.at(redis.get(lock_key).to_i) > Time.now
+            delete_lock
+            acquire_lock
+          else
+            false
+          end
+        end
+
+        def delete_lock
+          redis.del(lock_key)
+        end
+
+        def enqueue(override = false)
+          count_apps = 0
+          count_not  = 0
+          locked     = false
+
+          unless override
+            count_apps = queued_count
+            if count_apps == 0
+              locked     = acquire_lock
+              return unless locked
+              count_apps = queued_count
+            end
+            count_not  = Resque.aps_notification_count_for_application(name)
+          end
+
+          if count_apps <= 0 || (count_apps < Resque.aps_application_job_limit && (count_not > Resque.aps_queue_size_upper && count_not % (Resque.aps_queue_size_upper / 10) == 0))
+            enqueue(Resque::Plugins::Aps::Application, name)
+            redis.incr(Resque.aps_application_queued_key(name))
+            delete_lock if locked
+          end
+        end
+
+        def dequeue
+          redis.decr(Resque.aps_application_queued_key(name))
+        end
+
       end
     end
   end